Reverted the use of html in the dictionary view
In F18, WebKit and Gst have a conlict and the activity crash when
used together. I reverted the use of WebKit to display the html
on the new English dictionary. We loose the rich text.
Only bold and italic is used, but would be nice if we can parse the
html and insert the tags to get them back.