Create and use a sql index
A big dictionary, like the English (180K definitions)
take a long time to load. We can put it on a sql database,
make initialization faster and the query a little slower.
The use of the index is not mandaory, if is available is useed.
To create the index, in needed execute the method create_sql_index()